Fixed GET command
Due to self.model being written as a string, that meant all the
self.model checks were actually stealthily failing.

Hence get was actually not differentiating between model 1, 2, etc. Once
this was fixed, there was also an error in the handling of get when
model == 2, i.e. it crashed when there was no key. This has been fixed.